web_scraping_project_files/scraper.py

# This file will contain the code we discussed earlier for scraping, saving, and emailing job listings.
import json
import os
import schedule
import smtplib
import time
from selenium import webdriver
from bs4 import BeautifulSoup
from email.mime.multipart import MIMEMultipart
from email.mime.text import MIMEText
from email.mime.base import MIMEBase
from email import encoders
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# Selenium setup with headless browser
def setup_driver():
    options = webdriver.ChromeOptions()
    options.add_argument('headless')
    driver_path = os.getenv('WEBDRIVER_PATH')
    driver = webdriver.Chrome(driver_path, options=options)
    return driver

# Define other functions here...

# Scheduled scraping
def scheduled_scraping():
    # Define scraping logic here...
    pass

# Schedule the scraping to run once every day at a time specified in the .env
schedule_time = os.getenv('SCHEDULE_TIME', '10:00')
schedule.every().day.at(schedule_time).do(scheduled_scraping)

# Run the scheduled tasks
while True:
    schedule.run_pending()
    time.sleep(1)
