web_scraping_project_files/requirements.txt

selenium==3.141.0
beautifulsoup4==4.9.3
schedule==1.1.0
python-dotenv==0.19.1
